LayoutTests/inspector/debugger/scripts-panel-expected.txt

Tests that scripts panel UI elements work as intended.


Running: testInitialLoad
Dumping ScriptsNavigator 'Scripts' tab:
  (no domain)
    bar.js
    baz.js
    foo.js
    foobar.js
Dumping ScriptsNavigator 'Content scripts' tab:

Source requested for baz.js

Running: testReset
Source requested for baz.js
Dumping ScriptsNavigator 'Scripts' tab:
  (no domain)
    bar.js
    baz.js
    foo.js
Dumping ScriptsNavigator 'Content scripts' tab:

Dumping ScriptsNavigator 'Scripts' tab:
Dumping ScriptsNavigator 'Content scripts' tab:

Dumping ScriptsNavigator 'Scripts' tab:
  (no domain)
    bar.js
Dumping ScriptsNavigator 'Content scripts' tab:


Running: testDebuggerUISourceCodeAddedAndRemoved
Dumping ScriptsNavigator 'Scripts' tab:
  (no domain)
    foo.js
Dumping ScriptsNavigator 'Content scripts' tab:

Dumping ScriptsNavigator 'Scripts' tab:
  (no domain)
    foo.js
    source.js
Dumping ScriptsNavigator 'Content scripts' tab:
